Move event handling fns out of kas-core::event::cx::platform
