+++ crates/kas-core/src/event/cx/platform.rs
@@ -7,7 +7,7 @@
 
 use super::*;
 use crate::theme::ThemeSize;
-use crate::{Tile, window::Window};
+use crate::window::Window;
 
 #[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
 #[cfg_attr(docsrs, doc(cfg(internal_doc)))]
@@ -105,204 +105,5 @@ impl EventState {
             scroll: Scroll::None,
         };
         f(&mut cx);
-    }
-
-    /// Handle all pending items before event loop sleeps
-    pub(crate) fn flush_pending<'a, A>(
-        &'a mut self,
-        runner: &'a mut dyn RunnerT,
-        window: &'a dyn WindowDataErased,
-        messages: &'a mut MessageStack,
-        win: &mut Window<A>,
-        data: &A,
-    ) -> Action {
-        self.with(runner, window, messages, |cx| {
-            while let Some((id, wid)) = cx.popup_removed.pop() {
-                cx.send_event(win.as_node(data), id, Event::PopupClosed(wid));
-            }
-
-            cx.mouse_handle_pending(win, data);
-            cx.touch_handle_pending(win, data);
-
-            if let Some(id) = cx.pending_update.take() {
-                win.as_node(data).find_node(&id, |node| cx.update(node));
-            }
-
-            match std::mem::take(&mut cx.pending_nav_focus) {
-                PendingNavFocus::None => (),
-                PendingNavFocus::Set { target, source } => {
-                    cx.set_nav_focus_impl(win.as_node(data), target, source)
-                }
-                PendingNavFocus::Next {
-                    target,
-                    reverse,
-                    source,
-                } => cx.next_nav_focus_impl(win.as_node(data), target, reverse, source),
-            }
-
-            // Update sel focus after nav focus:
-            if let Some(pending) = cx.pending_sel_focus.take() {
-                cx.set_sel_focus(cx.window, win.as_node(data), pending);
-            }
-
-            while let Some((id, msg)) = cx.send_queue.pop_front() {
-                cx.send_or_replay(win.as_node(data), id, msg);
-            }
-
-            // Poll futures. TODO(opt): this does not need to happen so often,
-            // but just in frame_update is insufficient.
-            cx.poll_futures(win.as_node(data));
-
-            // Finally, clear the region_moved flag (mouse and touch sub-systems handle this).
-            if cx.action.contains(Action::REGION_MOVED) {
-                cx.action.remove(Action::REGION_MOVED);
-                cx.action.insert(Action::REDRAW);
-            }
-        });
-
-        if let Some(icon) = self.mouse.update_cursor_icon() {
-            window.set_cursor_icon(icon);
-        }
-
-        std::mem::take(&mut self.action)
-    }
-
-    /// Window has been closed: clean up state
-    pub(crate) fn suspended(&mut self, runner: &mut dyn RunnerT) {
-        while !self.popups.is_empty() {
-            let id = self.close_popup(self.popups.len() - 1);
-            runner.close_window(id);
-        }
-    }
-}
-
-#[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
-#[cfg_attr(docsrs, doc(cfg(internal_doc)))]
-impl<'a> EventCx<'a> {
-    /// Handle a winit `WindowEvent`.
-    ///
-    /// Note that some event types are not handled, since for these
-    /// events the graphics backend must take direct action anyway:
-    /// `Resized(size)`, `RedrawRequested`, `HiDpiFactorChanged(factor)`.
-    pub(crate) fn handle_winit<A>(
-        &mut self,
-        win: &mut Window<A>,
-        data: &A,
-        event: winit::event::WindowEvent,
-    ) {
-        use winit::event::WindowEvent::*;
-
-        match event {
-            CloseRequested => self.action(win.id(), Action::CLOSE),
-            /* Not yet supported: see #98
-            DroppedFile(path) => ,
-            HoveredFile(path) => ,
-            HoveredFileCancelled => ,
-            */
-            Focused(state) => {
-                self.window_has_focus = state;
-                if state {
-                    // Required to restart theme animations
-                    self.action(win.id(), Action::REDRAW);
-                } else {
-                    // Window focus lost: close all popups
-                    while let Some(id) = self.popups.last().map(|state| state.id) {
-                        self.close_window(id);
-                    }
-                }
-            }
-            KeyboardInput {
-                mut event,
-                is_synthetic,
-                ..
-            } => {
-                let state = event.state;
-                let physical_key = event.physical_key;
-                let logical_key = event.logical_key.clone();
-
-                if let Some(id) = self.key_focus() {
-                    // TODO(winit): https://github.com/rust-windowing/winit/issues/3038
-                    let mut mods = self.modifiers;
-                    mods.remove(ModifiersState::SHIFT);
-                    if !mods.is_empty()
-                        || event
-                            .text
-                            .as_ref()
-                            .and_then(|t| t.chars().next())
-                            .map(|c| c.is_control())
-                            .unwrap_or(false)
-                    {
-                        event.text = None;
-                    }
-
-                    if self.send_event(win.as_node(data), id, Event::Key(&event, is_synthetic)) {
-                        return;
-                    }
-                }
-
-                if state == ElementState::Pressed && !is_synthetic {
-                    self.start_key_event(win.as_node(data), logical_key, physical_key);
-                } else if state == ElementState::Released
-                    && let Some(id) = self.key_depress.remove(&physical_key)
-                {
-                    self.redraw(id);
-                }
-            }
-            ModifiersChanged(modifiers) => {
-                let state = modifiers.state();
-                if state.alt_key() != self.modifiers.alt_key() {
-                    // This controls drawing of access key indicators
-                    self.action(win.id(), Action::REDRAW);
-                }
-                self.modifiers = state;
-            }
-            Ime(winit::event::Ime::Enabled) => {
-                // We expect self.ime.is_some(), but it's possible that the request is outdated
-                if self.ime.is_some()
-                    && let Some(id) = self.sel_focus.clone()
-                {
-                    self.send_event(win.as_node(data), id, Event::ImeFocus);
-                }
-            }
-            Ime(winit::event::Ime::Disabled) => {
-                // We can only assume that this is received due to us disabling
-                // IME if self.old_ime_target is set, and is otherwise due to an
-                // external cause.
-                let mut target = self.old_ime_target.take();
-                if target.is_none() && self.ime.is_some() {
-                    target = self.sel_focus.clone();
-                    self.ime = None;
-                    self.ime_cursor_area = Rect::ZERO;
-                }
-                if let Some(id) = target {
-                    self.send_event(win.as_node(data), id, Event::LostImeFocus);
-                }
-            }
-            Ime(winit::event::Ime::Preedit(text, cursor)) => {
-                if self.ime.is_some()
-                    && let Some(id) = self.sel_focus.clone()
-                {
-                    self.send_event(win.as_node(data), id, Event::ImePreedit(&text, cursor));
-                }
-            }
-            Ime(winit::event::Ime::Commit(text)) => {
-                if self.ime.is_some()
-                    && let Some(id) = self.sel_focus.clone()
-                {
-                    self.send_event(win.as_node(data), id, Event::ImeCommit(&text));
-                }
-            }
-            CursorMoved { position, .. } => self.handle_cursor_moved(win, data, position.into()),
-            CursorEntered { .. } => self.handle_cursor_entered(),
-            CursorLeft { .. } => self.handle_cursor_left(win.as_node(data)),
-            MouseWheel { delta, .. } => self.handle_mouse_wheel(win.as_node(data), delta),
-            MouseInput { state, button, .. } => {
-                self.handle_mouse_input(win.as_node(data), state, button)
-            }
-            // TouchpadPressure { pressure: f32, stage: i64, },
-            // AxisMotion { axis: AxisId, value: f64, },
-            Touch(touch) => self.handle_touch_event(win, data, touch),
-            _ => (),
-        }
     }
 }
